+++ packages/xrpl/test/models/AMMCreate.test.ts
@@ -1,5 +1,11 @@
 import { validateAMMCreate } from '../../src/models/transactions/AMMCreate'
-import { assertTxIsValid, assertTxValidationError } from '../testUtils'
+import {
+  assertTxIsValid,
+  assertTxValidationError,
+  MPT_ISSUANCE_ID_1,
+  MPT_ISSUANCE_ID_2,
+  MPTID_LENGTH,
+} from '../testUtils'
 
 const assertValid = (tx: any): void => assertTxIsValid(tx, validateAMMCreate)
 const assertInvalid = (tx: any, message: string): void =>
@@ -77,6 +83,58 @@ describe('AMMCreate', function () {
   it(`throws when TradingFee is a negative number`, function () {
     ammCreate.TradingFee = -1
     const errorMessage = 'AMMCreate: TradingFee must be between 0 and 1000'
+    assertInvalid(ammCreate, errorMessage)
+  })
+
+  // MPT-related tests
+  it(`verifies valid AMMCreate with two MPT assets`, function () {
+    ammCreate.Amount = {
+      mpt_issuance_id: MPT_ISSUANCE_ID_1,
+      value: '250',
+    }
+    ammCreate.Amount2 = {
+      mpt_issuance_id: MPT_ISSUANCE_ID_2,
+      value: '250',
+    }
+    assertValid(ammCreate)
+  })
+
+  it(`throws w/ Amount mpt_issuance_id contains non-hex characters`, function () {
+    ammCreate.Amount = {
+      mpt_issuance_id: 'Z'.repeat(MPTID_LENGTH),
+      value: '250',
+    }
+    ammCreate.Amount2 = {
+      mpt_issuance_id: MPT_ISSUANCE_ID_2,
+      value: '250',
+    }
+    const errorMessage = 'AMMCreate: Amount must be an Amount'
+    assertInvalid(ammCreate, errorMessage)
+  })
+
+  it(`throws w/ Amount mpt_issuance_id too short`, function () {
+    ammCreate.Amount = {
+      mpt_issuance_id: 'A'.repeat(MPTID_LENGTH - 1),
+      value: '250',
+    }
+    ammCreate.Amount2 = {
+      mpt_issuance_id: MPT_ISSUANCE_ID_2,
+      value: '250',
+    }
+    const errorMessage = 'AMMCreate: Amount must be an Amount'
+    assertInvalid(ammCreate, errorMessage)
+  })
+
+  it(`throws w/ Amount mpt_issuance_id too long`, function () {
+    ammCreate.Amount = {
+      mpt_issuance_id: 'A'.repeat(MPTID_LENGTH + 1),
+      value: '250',
+    }
+    ammCreate.Amount2 = {
+      mpt_issuance_id: MPT_ISSUANCE_ID_2,
+      value: '250',
+    }
+    const errorMessage = 'AMMCreate: Amount must be an Amount'
     assertInvalid(ammCreate, errorMessage)
   })
 })